v14

+++ 2.1/main.c
@@ -6,8 +6,8 @@
 int main(void){
 	char str1[16] = "Hello!";
 	char* str2 = "Another string";
-	printf("Length of str1: %u\n", (unsigned int)mystrlen(str1));
-	printf("Length of str2: %u\n", (unsigned int)mystrlen(str2));
-	printf("Custom length: %u\n", (unsigned int)mystrlen(str2 + 7));
+	printf("Length of str1: %lu\n", mystrlen(str1));
+	printf("Length of str2: %lu\n", mystrlen(str2));
+	printf("Custom length: %lu\n", mystrlen(str2 + 7));
 	return 0;
 }
